requirements.txt
google-auth
google-auth-oauthlib
google-auth-httplib2
google-cloud-speech
google-cloud-texttospeech
grpcio
bokeh
streamlit-bokeh-events
openai==0.27.0
